Core: Move vanadiel_clock to it's own file
